refactor: move stripNullish into src/common and reuse in Kostenersatz
Konsolidiert den bisherigen lokalen `removeUndefined`-Helper aus
`useKostenersatzMutations.ts` und das in PR #499 eingeführte `stripNullish`
aus `einsatzDefaults.ts` zu einem gemeinsamen Modul `src/common/stripNullish.ts`.
Verhalten unverändert (entfernt undefined- und null-Felder vor Firestore-Writes).
