+++ src/components/FirecallItems/einsatzDefaults.test.ts
@@ -4,7 +4,6 @@ import {
   DEFAULT_EINSATZ_FW,
   DEFAULT_EINSATZ_GROUP,
   resetEinsatzToManual,
-  stripNullish,
 } from './einsatzDefaults';
 import { Firecall } from '../firebase/firestore';
 
@@ -72,32 +71,5 @@ describe('resetEinsatzToManual', () => {
 
     expect(reset.group).toBe('custom-group');
     expect(reset.fw).toBe('Custom FW');
-  });
-});
-
-describe('stripNullish', () => {
-  it('removes undefined and null values', () => {
-    const input = {
-      a: 'hello',
-      b: undefined,
-      c: null,
-      d: 0,
-      e: '',
-      f: false,
-    };
-
-    const result = stripNullish(input);
-
-    expect(result).toEqual({ a: 'hello', d: 0, e: '', f: false });
-    expect('b' in result).toBe(false);
-    expect('c' in result).toBe(false);
-  });
-
-  it('preserves nested objects as-is', () => {
-    const nested = { foo: 'bar' };
-    const result = stripNullish({ nested, skip: undefined });
-
-    expect(result.nested).toBe(nested);
-    expect('skip' in result).toBe(false);
   });
 });
