+++ src/components/FirecallItems/einsatzDefaults.ts
@@ -1,17 +1,6 @@
 import { formatTimestamp } from '../../common/time-format';
 import { Firecall } from '../firebase/firestore';
 
-export function stripNullish<T extends Record<string, any>>(obj: T): Partial<T> {
-  const result: Partial<T> = {};
-  for (const key of Object.keys(obj) as (keyof T)[]) {
-    const value = obj[key];
-    if (value !== undefined && value !== null) {
-      result[key] = value;
-    }
-  }
-  return result;
-}
-
 export const DEFAULT_EINSATZ_GROUP = 'ffnd';
 export const DEFAULT_EINSATZ_FW = 'Neusiedl am See';
